requirements.txt
Django==3.1.4
django-filter==2.4.0
djangorestframework==3.12.2
Markdown==3.3.3
 flake8==3.8.4